[Rearch]Domain層とInfra層を追加し依存関係、データの流れをわかりやすくした
