+++ app/handler/twitter.go
@@ -31,21 +31,6 @@ func (th TwitterHandlerImpl) HandleTwitterPostWebhook(ctx *gin.Context) {
 		ctx.JSON(http.StatusBadRequest, err)
 	}
 
-	res := usecase.PostTwitterWebhookResponse{}
+	res := th.TwitterUseCase.PostAutoReplyResponse(req)
 	ctx.JSON(http.StatusOK, res)
-
-	// if len(req.TweetCreateEvents) < 1 || req.UserID == req.TweetCreateEvents[0].User.IDStr {
-	// 	return
-	// }
-
-	// 自動でリプライを返す
-	// api := twitter.GetTwitterApi()
-	// params := url.Values{}
-	// params.Set("in_reply_to_status_id", req.TweetCreateEvents[0].TweetIDStr)
-	// _, err := api.PostTweet("@"+req.TweetCreateEvents[0].User.ScreenName+" Hello World", params)
-	// if err != nil {
-	// 	ctx.JSON(http.StatusBadRequest, err)
-	// } else {
-	// 	ctx.Status(200)
-	// }
 }
